+++ src/main/java/kitchenpos/application/OrderService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import kitchenpos.dao.OrderLineItemDao;
-import kitchenpos.domain.MenuRepository;
+import kitchenpos.domain.menu.MenuRepository;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderRepository;
